Updated implementation

+++ jext.linalg/src/test/java/jext/math/linear/TestCoords.java
@@ -14,18 +14,18 @@ public class TestCoords extends Assertions {
 
     @Test
     void testUnion() {
-        Coords c1 = new Coords(new long[]{1,3,5,7});
-        Coords c2 = new Coords(new long[]{2,4,6,7,8});
-        Coords c3 = new Coords(new long[]{1,2,3,4,5,6,7,8});
+        Coords c1 = new Coords(new int[]{1,3,5,7});
+        Coords c2 = new Coords(new int[]{2,4,6,7,8});
+        Coords c3 = new Coords(new int[]{1,2,3,4,5,6,7,8});
         Coords u = c1.union(c2);
         assertEquals(c3, u);
     }
 
     @Test
     void testIntersection() {
-        Coords c1 = new Coords(new long[]{1,3,5,7});
-        Coords c2 = new Coords(new long[]{2,4,6,7,8});
-        Coords c4 = new Coords(new long[]{7});
+        Coords c1 = new Coords(new int[]{1,3,5,7});
+        Coords c2 = new Coords(new int[]{2,4,6,7,8});
+        Coords c4 = new Coords(new int[]{7});
         Coords i = c1.intersection(c2);
         assertEquals(c4, i);
     }
